requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
decorator==5.2.1
filelock==3.16.1
Flask==2.0.1
fonttools==4.57.0
fsspec==2025.3.0
future==1.0.0
geocoder==1.38.1
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.8.0.74
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==7.0.0
py-cpuinfo==9.0.0
pygame==2.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pywhatkit==5.4
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
requests==2.31.0
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
thop==0.1.1.post2209072238
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing-extensions==4.13.2
tzdata==2025.2
ultralytics==8.0.196
urllib3==2.2.3
werkzeug==2.3.7
zipp==3.20.2
twilio==8.10.0